+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -83,5 +83,9 @@ public enum NoteTaskIdentifier
     
     [Description("master_funding_agreement")]
     [DisplayDescription("Master funding agreement")]
-    MasterFundingAgreement = 20
+    MasterFundingAgreement = 20,
+
+    [Description("confirm_incoming_trust_ceo_contact")]
+    [DisplayDescription("Confirm the incoming trust ceo contact")]
+    ConfirmIncomingTrustCeoContact = 21
 }